refactor: update metadata
