+++ src/common/meta/src/ddl/create_logical_tables.rs
@@ -15,36 +15,29 @@
 mod check;
 mod metadata;
 mod region_request;
-
-use std::ops::Deref;
+mod update_metadata;
 
 use api::v1::CreateTableExpr;
 use async_trait::async_trait;
 use common_procedure::error::{FromJsonSnafu, Result as ProcedureResult, ToJsonSnafu};
 use common_procedure::{Context as ProcedureContext, LockKey, Procedure, Status};
-use common_telemetry::{info, warn};
+use common_telemetry::warn;
 use futures_util::future::join_all;
-use itertools::Itertools;
 use serde::{Deserialize, Serialize};
-use snafu::{ensure, OptionExt, ResultExt};
+use snafu::{ensure, ResultExt};
 use store_api::metadata::ColumnMetadata;
 use store_api::metric_engine_consts::ALTER_PHYSICAL_EXTENSION_KEY;
 use store_api::storage::{RegionId, RegionNumber};
 use strum::AsRefStr;
 use table::metadata::{RawTableInfo, TableId};
 
-use crate::cache_invalidator::Context;
 use crate::ddl::utils::{add_peer_context_if_needed, handle_retry_error};
-use crate::ddl::{physical_table_metadata, DdlContext};
-use crate::error::{DecodeJsonSnafu, MetadataCorruptionSnafu, Result, TableInfoNotFoundSnafu};
-use crate::instruction::CacheIdent;
-use crate::key::table_info::TableInfoValue;
+use crate::ddl::DdlContext;
+use crate::error::{DecodeJsonSnafu, MetadataCorruptionSnafu, Result};
 use crate::key::table_route::TableRouteValue;
-use crate::key::DeserializedValueWithBytes;
 use crate::lock_key::{CatalogLock, SchemaLock, TableLock, TableNameLock};
 use crate::rpc::ddl::CreateTableTask;
 use crate::rpc::router::{find_leaders, RegionRoute};
-use crate::table_name::TableName;
 use crate::{metrics, ClusterId};
 
 pub struct CreateLogicalTablesProcedure {
@@ -138,86 +131,8 @@ impl CreateLogicalTablesProcedure {
     /// Abort(not-retry):
     /// - Failed to create table metadata.
     pub async fn on_create_metadata(&mut self) -> Result<Status> {
-        let manager = &self.context.table_metadata_manager;
-        let remaining_tasks = self.data.remaining_tasks();
-        let num_tables = remaining_tasks.len();
-
-        if num_tables > 0 {
-            let chunk_size = manager.create_logical_tables_metadata_chunk_size();
-            if num_tables > chunk_size {
-                let chunks = remaining_tasks
-                    .into_iter()
-                    .chunks(chunk_size)
-                    .into_iter()
-                    .map(|chunk| chunk.collect::<Vec<_>>())
-                    .collect::<Vec<_>>();
-                for chunk in chunks {
-                    manager.create_logical_tables_metadata(chunk).await?;
-                }
-            } else {
-                manager
-                    .create_logical_tables_metadata(remaining_tasks)
-                    .await?;
-            }
-        }
-
-        // The `table_id` MUST be collected after the [Prepare::Prepare],
-        // ensures the all `table_id`s have been allocated.
-        let table_ids = self
-            .data
-            .tasks
-            .iter()
-            .map(|task| task.table_info.ident.table_id)
-            .collect::<Vec<_>>();
-
-        if !self.data.physical_columns.is_empty() {
-            // fetch old physical table's info
-            let physical_table_info = self
-                .context
-                .table_metadata_manager
-                .table_info_manager()
-                .get(self.data.physical_table_id)
-                .await?
-                .with_context(|| TableInfoNotFoundSnafu {
-                    table: format!("table id - {}", self.data.physical_table_id),
-                })?;
-
-            // generate new table info
-            let new_table_info = self
-                .data
-                .build_new_physical_table_info(&physical_table_info);
-
-            let physical_table_name = TableName::new(
-                &new_table_info.catalog_name,
-                &new_table_info.schema_name,
-                &new_table_info.name,
-            );
-
-            // update physical table's metadata
-            self.context
-                .table_metadata_manager
-                .update_table_info(physical_table_info, new_table_info)
-                .await?;
-
-            // invalid table cache
-            self.context
-                .cache_invalidator
-                .invalidate(
-                    &Context::default(),
-                    vec![
-                        CacheIdent::TableId(self.data.physical_table_id),
-                        CacheIdent::TableName(physical_table_name),
-                    ],
-                )
-                .await?;
-        } else {
-            warn!("No physical columns found, leaving the physical table's schema unchanged");
-        }
-
-        info!(
-            "Created {num_tables} tables {table_ids:?} metadata for physical table {}",
-            self.data.physical_table_id
-        );
+        self.update_physical_table_metadata().await?;
+        let table_ids = self.create_logical_tables_metadata().await?;
 
         Ok(Status::done_with_output(table_ids))
     }
@@ -369,21 +284,6 @@ impl CreateTablesData {
                 }
             })
             .collect::<Vec<_>>()
-    }
-
-    /// Generate the new physical table info.
-    ///
-    /// This method will consumes the physical columns.
-    fn build_new_physical_table_info(
-        &mut self,
-        old_table_info: &DeserializedValueWithBytes<TableInfoValue>,
-    ) -> RawTableInfo {
-        let raw_table_info = old_table_info.deref().table_info.clone();
-
-        physical_table_metadata::build_new_physical_table_info(
-            raw_table_info,
-            &self.physical_columns,
-        )
     }
 }
